+++ Arduino/libraries/AHRS/AHRS.h
@@ -66,6 +66,6 @@ void rotMatToQuat(Eigen::MatrixXf& Rot, float q[4]);
 
 float median(float array[]);
 float mean(float array[]);
-uint8_t checksum(uint8_t array[], int first, int last);
+uint16_t checksum(uint8_t array[], int first, int last);
 
 #endif
